lib: signed/unsigned update for crypto

+++ lib/crypto/aes128.h
@@ -1,13 +1,13 @@
-#include <bits/ints.h>
+#include <bits/types.h>
 
 struct aes128 {
 	uint32_t W[44];  /* 11 round keys, 4x4 each */
 };
 
-void aes128_init(struct aes128* ctx, const uint8_t key[16]);
-void aes128_decrypt(struct aes128* ctx, uint8_t blk[16]);
-void aes128_encrypt(struct aes128* ctx, uint8_t blk[16]);
+void aes128_init(struct aes128* ctx, const byte key[16]);
+void aes128_decrypt(struct aes128* ctx, byte blk[16]);
+void aes128_encrypt(struct aes128* ctx, byte blk[16]);
 void aes128_fini(struct aes128* ctx);
 
-void aes128_wrap(uint8_t key[16], void* buf, int len);
-void aes128_unwrap(uint8_t key[16], void* buf, int len);
+void aes128_wrap(byte key[16], void* buf, ulong len);
+void aes128_unwrap(byte key[16], void* buf, ulong len);
