+++ lib/crypto/aes128_wrap.c
@@ -13,14 +13,14 @@ static uint64_t wrapmask(int n)
 #endif
 }
 
-void aes128_wrap(uint8_t key[16], void* buf, int len)
+void aes128_wrap(byte key[16], void* buf, ulong len)
 {
 	struct aes128 ae;
 
-	int n = len / 8 - 1;
+	ulong n = len / 8 - 1;
 	uint64_t* R = buf;
-	uint8_t B[16];
-	int i, j;
+	byte B[16];
+	uint i, j;
 
 	aes128_init(&ae, key);
